restapi/requirements.txt
Cython
email-validator
pydantic --no-binary pydantic
python-dotenv
python-multipart
python-slugify
psycopg2-binary
SQLAlchemy
databases[postgresql]
alembic
fastapi
uvicorn[standard]
bcrypt
orjson
jinja2
redis
aiofiles
aiosmtplib
requests
Authlib
httpx
itsdangerous
Pillow
phonenumbers
fastapi-jwt-auth[asymmetric]
user-agents
dhash
pytz

# testing
pytest
pytest-cov
pytest-asyncio
